move noteOptions to values directory
