+++ client/components/keyboard-full.tsx
@@ -5,24 +5,10 @@
 
 import Scrollbar from '@/components/common/scrollbar';
 import Octave from '@/components/octave';
+import { noteOptions } from '@/values/settingsOptions';
 
 // create an array of numbers 1-7
 const octaveNums = Array.from({ length: 7 }, (_, index) => index + 1);
-
-export const noteOptions = [
-	'C',
-	'D♭',
-	'D',
-	'E♭',
-	'E',
-	'F',
-	'G♭',
-	'G',
-	'A♭',
-	'A',
-	'B♭',
-	'B',
-];
 
 export default function KeyboardFull(): JSX.Element {
 	return (
