+++ client/values/settingsOptions.ts
@@ -1,0 +1,14 @@
+export const noteOptions = [
+	'C',
+	'D♭',
+	'D',
+	'E♭',
+	'E',
+	'F',
+	'G♭',
+	'G',
+	'A♭',
+	'A',
+	'B♭',
+	'B',
+];
